requirements.txt
apache-airflow-providers-mongo==3.0.0
apache-airflow-providers-docker==3.8.1